RequestRefresh back to a signal for GetConfigurationAsync (#3087)
* RequestRefresh back to a signal

Further reverting the change to RequestRefresh from #2780
This is still too much of a behavioral change. RequestRefresh goes back
to its historical function, signalling that the next call to
GetConfigurationAsync should request data. This is incorporated with
the background thread change to instead do the request as a blocking
operation if from a RequestRefresh

Part of #3082

* Add time based testing to ConfigurationManagerTests
through FakeTimeProvider